feat: migrate to spa
