--- renderer/pages/_document.tsx
@@ -1,54 +1,0 @@
-import React from 'react'
-import Document, { Head, Main, NextScript } from 'next/document'
-import { ServerStyleSheets } from '@material-ui/styles'
-import { theme } from '../lib/theme'
-import { initializeFonts } from '../lib/fonts'
-
-class MyDocument extends Document {
-  componentDidMount(): void {
-    initializeFonts()
-  }
-
-  render(): React.ReactElement {
-    return (
-      <html lang="en" dir="ltr">
-        <Head>
-          <meta charSet="utf-8" />
-          <meta
-            name="viewport"
-            content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no"
-          />
-          <meta name="theme-color" content={theme.palette.primary.main} />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </html>
-    )
-  }
-}
-
-MyDocument.getInitialProps = async (ctx) => {
-  const sheets = new ServerStyleSheets()
-  const originalRenderPage = ctx.renderPage
-
-  ctx.renderPage = () =>
-    originalRenderPage({
-      enhanceApp: (App) => (props) => sheets.collect(<App {...props} />),
-    })
-
-  const initialProps = await Document.getInitialProps(ctx)
-
-  return {
-    ...initialProps,
-    styles: [
-      <React.Fragment key="styles">
-        {initialProps.styles}
-        {sheets.getStyleElement()}
-      </React.Fragment>,
-    ],
-  }
-}
-
-export default MyDocument
