+++ src/main/ipcs/type.d.ts
