+++ src/renderer/components/atoms/HurryAnimationText.tsx
