+++ src/renderer/components/atoms/TimerLabel.tsx
