+++ src/renderer/components/atoms/TimerValue.tsx
