+++ src/renderer/components/molecules/AgendaSkinCircle.tsx
