+++ src/renderer/components/molecules/AgendaSkinList.tsx
