+++ src/renderer/components/molecules/TimerInfo.tsx
