+++ src/renderer/i18n.ts
