+++ src/renderer/locales/en.ts
