+++ src/renderer/locales/ja.ts
