+++ src/renderer/modules/ipc/type.d.ts
