+++ src/renderer/modules/setting/reducer.ts
