+++ src/renderer/modules/setting/type.d.ts
