+++ src/renderer/modules/timer/reducer.spec.ts
