+++ src/renderer/modules/timer/reducer.ts
