+++ src/renderer/modules/timer/type.d.ts
