+++ src/renderer/modules/timer/util.spec.ts
