+++ src/renderer/modules/timer/util.ts
