+++ src/renderer/modules/type.d.ts
