+++ src/renderer/modules/updater/reducer.ts
