+++ src/renderer/public/images/logo.png
