+++ src/renderer/reducers.ts
